split ba classes into separate files
